requirements.txt
wheel
tensorflow_datasets
pandas
matplotlib
ipykernel
